Create README.md
This program calculates salary of employee or an employer in a month . It assumes the total number of working hours is 20. Also, it assumes an employee is paid $160 every hour and an employer is paid $1000 every hour. A worker also enjoys 10% benefit on the salary, if the person does not absent him or herself from work in a month. It also makes deduction of salary based on hours lost however the employer is not entitled to any deduction.